feat/authenticator: implement recovery for app authentication
Make sure that if the app authentication operation is failed, it
doesn't lock out app/user from subsequent attempts to authenticate
the same app
